corriendo temp

+++ services/whisper.js
@@ -33,9 +33,25 @@ const voiceToText = async (path) => {
 		);
 
 		//delete all files in folder tmp
-		fs.readdirSync('./tmp').forEach((file) => {
-			fs.unlinkSync(`./tmp/${file}`);
-		});
+		try {
+			// Verifica si el directorio existe antes de intentar leerlo
+			if (fs.existsSync('./tmp')) {
+				// Lee el contenido del directorio y elimina cada archivo
+				fs.readdirSync('./tmp').forEach((file) => {
+					try {
+						fs.unlinkSync(`./tmp/${file}`);
+					} catch (err) {
+						console.error(`Error al eliminar el archivo: ${file}`, err);
+					}
+				});
+			} else {
+				console.log(
+					'El directorio ./tmp no existe, continuando con el proceso.'
+				);
+			}
+		} catch (err) {
+			console.error('Error al leer el directorio ./tmp:', err);
+		}
 
 		return response.data.text;
 	} catch (err) {
